requirements.txt
beautifulsoup4==4.9.1
notebook==6.1.1
nltk==3.5
mypy==0.782
graphviz==0.14.1
pydot==1.4.1
python-benedict==0.18.1